+++ ext/com_dotnet/com_saproxy.c
@@ -167,11 +167,11 @@ static zval *saproxy_read_dimension(zval *object, zval *offset, int type, zval *
 		/* copy indices from proxy */
 		for (i = 0; i < dims; i++) {
 			convert_to_long(&proxy->indices[i]);
-			indices[i] = Z_LVAL(proxy->indices[i]);
+			indices[i] = (LONG)Z_LVAL(proxy->indices[i]);
 		}
 
 		/* add user-supplied index */
-		indices[dims-1] = Z_LVAL_P(offset);
+		indices[dims-1] = (LONG)Z_LVAL_P(offset);
 
 		/* now fetch the value */
 		if (FAILED(SafeArrayGetVartype(sa, &vt)) || vt == VT_EMPTY) {
@@ -241,12 +241,12 @@ static void saproxy_write_dimension(zval *object, zval *offset, zval *value TSRM
 		/* copy indices from proxy */
 		for (i = 0; i < dims; i++) {
 			convert_to_long(&proxy->indices[i]);
-			indices[i] = Z_LVAL(proxy->indices[i]);
+			indices[i] = (LONG)Z_LVAL(proxy->indices[i]);
 		}
 
 		/* add user-supplied index */
 		convert_to_long(offset);
-		indices[dims-1] = Z_LVAL_P(offset);
+		indices[dims-1] = (LONG)Z_LVAL_P(offset);
 
 		if (FAILED(SafeArrayGetVartype(V_ARRAY(&proxy->obj->v), &vt)) || vt == VT_EMPTY) {
 			vt = V_VT(&proxy->obj->v) & ~VT_ARRAY;
@@ -555,7 +555,7 @@ zend_object_iterator *php_com_saproxy_iter_get(zend_class_entry *ce, zval *objec
 	I->indices = safe_emalloc(proxy->dimensions + 1, sizeof(LONG), 0);
 	for (i = 0; i < proxy->dimensions; i++) {
 		convert_to_long(&proxy->indices[i]);
-		I->indices[i] = Z_LVAL(proxy->indices[i]);
+		I->indices[i] = (LONG)Z_LVAL(proxy->indices[i]);
 	}
 
 	SafeArrayGetLBound(V_ARRAY(&proxy->obj->v), proxy->dimensions, &I->imin);
