+++ ext/filter/sanitizing_filters.c
@@ -30,7 +30,7 @@ typedef unsigned long filter_map[256];
 static void php_filter_encode_html(zval *value, const unsigned char *chars)
 {
 	smart_str str = {0};
-	int len = Z_STRLEN_P(value);
+	size_t len = Z_STRLEN_P(value);
 	unsigned char *s = (unsigned char *)Z_STRVAL_P(value);
 	unsigned char *e = s + len;
 
@@ -147,7 +147,7 @@ static void filter_map_init(filter_map *map)
 
 static void filter_map_update(filter_map *map, int flag, const unsigned char *allowed_list)
 {
-	int l, i;
+	size_t l, i;
 
 	l = strlen((const char*)allowed_list);
 	for (i = 0; i < l; ++i) {
