+++ ext/standard/formatted_print.c
@@ -376,7 +376,7 @@ php_formatted_print(int param_count, int use_array, int format_offset TSRMLS_DC)
 	char *format, padding;
 	zend_string *result;
 	int always_sign;
-	int format_len;
+	size_t format_len;
 
 	if (zend_parse_parameters(param_count TSRMLS_CC, "+", &args, &argc) == FAILURE) {
 		return NULL;
