+++ ext/standard/php_http.h
@@ -25,9 +25,9 @@
 #include "zend_smart_str.h"
 
 PHPAPI int php_url_encode_hash_ex(HashTable *ht, smart_str *formstr,
-				const char *num_prefix, int num_prefix_len,
-				const char *key_prefix, int key_prefix_len,
-				const char *key_suffix, int key_suffix_len, 
+				const char *num_prefix, size_t num_prefix_len,
+				const char *key_prefix, size_t key_prefix_len,
+				const char *key_suffix, size_t key_suffix_len, 
 				zval *type, char *arg_sep, int enc_type TSRMLS_DC);
 #define php_url_encode_hash(ht, formstr)	php_url_encode_hash_ex((ht), (formstr), NULL, 0, NULL, 0, NULL, 0, NULL TSRMLS_CC)
