+++ ext/standard/type.c
@@ -157,7 +157,7 @@ PHP_FUNCTION(intval)
 #endif
 
 	RETVAL_ZVAL(num, 1, 0);
-	convert_to_long_base(return_value, base);
+	convert_to_long_base(return_value, (int)base);
 }
 /* }}} */
